consolidate dynamic groups state in separate module, fix dynamic groups view modifications
- Adds new dynamicGroups.ts module in @fiftyone/state and consolidates related state
- fixes falsey edge cases in group by expr
- filters GroupBy stage from modal and imavid requests
