+++ app/packages/state/src/recoil/groups.ts
@@ -6,13 +6,6 @@ import {
   groupSliceFragment,
   groupSliceFragment$key,
 } from "@fiftyone/relay";
-import {
-  DYNAMIC_GROUP_FIELDS,
-  EMBEDDED_DOCUMENT_FIELD,
-  GROUP,
-  LIST_FIELD,
-  Stage,
-} from "@fiftyone/utilities";
 import { get as getPath } from "lodash";
 import { VariablesOf } from "react-relay";
 import {
@@ -33,14 +26,17 @@ import {
 } from "./atoms";
 import { getBrowserStorageEffectForKey } from "./customEffects";
 import { dataset } from "./dataset";
+import {
+  imaVidLookerState,
+  isDynamicGroup,
+  isNestedDynamicGroup,
+  shouldRenderImaVidLooker,
+} from "./dynamicGroups";
 import { ModalSample, modalLooker, modalSample } from "./modal";
-import { dynamicGroupsViewMode } from "./options";
 import { RelayEnvironmentKey } from "./relay";
-import { fieldPaths } from "./schema";
 import { datasetName, parentMediaTypeSelector } from "./selectors";
-import { State } from "./types";
 import { mapSampleResponse } from "./utils";
-import { GROUP_BY_VIEW_STAGE, dynamicGroupViewQuery, view } from "./view";
+import * as viewAtoms from "./view";
 
 export const groupMediaIsCarouselVisibleSetting = atom<boolean>({
   key: "groupMediaIsCarouselVisibleSetting",
@@ -371,7 +367,9 @@ export const groupSamples = graphQLSelectorFamily<
       return {
         count,
         dataset: get(datasetName),
-        view: get(view),
+        view: get(viewAtoms.view).filter(
+          (stage) => stage._cls !== viewAtoms.GROUP_BY_VIEW_STAGE
+        ),
         filter: {
           group: {
             slice: get(groupSlice),
@@ -406,138 +404,6 @@ export const pcdSamples = selector({
         paginationData: false,
       })
     ),
-});
-
-export const groupByFieldValue = atom<string | null>({
-  key: "groupByFieldValue",
-  default: null,
-});
-
-export const dynamicGroupIndex = atom<number>({
-  key: "dynamicGroupIndex",
-  default: null,
-});
-
-export const dynamicGroupCurrentElementIndex = atom<number>({
-  key: "dynamicGroupCurrentElementIndex",
-  default: 1,
-});
-
-export const dynamicGroupParameters =
-  selector<State.DynamicGroupParameters | null>({
-    key: "dynamicGroupParameters",
-    get: ({ get }) => {
-      const viewArr = get(view);
-      if (!viewArr) return null;
-
-      const groupByViewStageNode = viewArr.find(
-        (view) => view._cls === GROUP_BY_VIEW_STAGE
-      );
-      if (!groupByViewStageNode) return null;
-
-      const isFlat = groupByViewStageNode.kwargs[2][1]; // third index is 'flat', we want it to be false for dynamic groups
-      if (isFlat) return null;
-
-      return {
-        groupBy: groupByViewStageNode.kwargs[0][1] as string, // first index is 'field_or_expr', which defines group-by
-        orderBy: groupByViewStageNode.kwargs[1][1] as string, // second index is 'order_by', which defines order-by
-      };
-    },
-  });
-
-export const isDynamicGroup = selector<boolean>({
-  key: "isDynamicGroup",
-  get: ({ get }) => {
-    return Boolean(get(dynamicGroupParameters));
-  },
-});
-
-export const isNonNestedDynamicGroup = selector<boolean>({
-  key: "isNonNestedDynamicGroup",
-  get: ({ get }) => {
-    return get(isDynamicGroup) && get(parentMediaTypeSelector) !== "group";
-  },
-});
-
-export const isNestedDynamicGroup = selector<boolean>({
-  key: "isNestedDynamicGroup",
-  get: ({ get }) => {
-    return get(isDynamicGroup) && get(hasGroupSlices);
-  },
-});
-
-export const imaVidStoreKey = selectorFamily<
-  string,
-  { modal: boolean; groupByFieldValue: string }
->({
-  key: "imaVidStoreKey",
-  get:
-    ({ modal, groupByFieldValue }) =>
-    ({ get }) => {
-      const { groupBy, orderBy } = get(dynamicGroupParameters);
-      const slice = get(currentSlice(modal)) ?? "UNSLICED";
-
-      return `$${groupBy}-${orderBy}-${groupByFieldValue}-${slice}`;
-    },
-});
-
-export const shouldRenderImaVidLooker = selector<boolean>({
-  key: "shouldRenderImaVidLooker",
-  get: ({ get }) => {
-    return get(isOrderedDynamicGroup) && get(dynamicGroupsViewMode) === "video";
-  },
-});
-
-export const isOrderedDynamicGroup = selector<boolean>({
-  key: "isOrderedDynamicGroup",
-  get: ({ get }) => {
-    const params = get(dynamicGroupParameters);
-    if (!params) return false;
-
-    const { orderBy } = params;
-    return Boolean(orderBy?.length);
-  },
-});
-
-export const dynamicGroupPageSelector = selectorFamily<
-  (
-    cursor: number,
-    pageSize: number
-  ) => {
-    filter: Record<string, never>;
-    after: string;
-    count: number;
-    dataset: string;
-    view: Stage[];
-  },
-  string
->({
-  key: "paginateDynamicGroupVariables",
-  get:
-    (groupByValue) =>
-    ({ get }) => {
-      const params = {
-        dataset: get(datasetName),
-        view: get(
-          dynamicGroupViewQuery({ groupByFieldValueExplicit: groupByValue })
-        ),
-      };
-
-      const filter = get(hasGroupSlices)
-        ? {
-            group: {
-              slice: get(groupSlice),
-            },
-          }
-        : {};
-
-      return (cursor: number, pageSize: number) => ({
-        ...params,
-        filter,
-        after: cursor ? String(cursor) : null,
-        count: pageSize,
-      });
-    },
 });
 
 export const activeModalSample = selector({
@@ -578,73 +444,7 @@ export const activeModalSidebarSample = selector({
   },
 });
 
-export const imaVidLookerState = atomFamily<any, string>({
-  key: "imaVidLookerState",
-  default: null,
-  effects: (key) => [
-    ({ setSelf, getPromise, onSet }) => {
-      let unsubscribe;
-
-      onSet((_newValue) => {
-        // note: resetRecoilState is not triggering `onSet` in effect,
-        // see https://github.com/facebookexperimental/Recoil/issues/2183
-        // replace with `useResetRecoileState` when fixed
-
-        // if (!isReset) {
-        //   throw new Error("cannot set ima-vid state directly");
-        // }
-        unsubscribe && unsubscribe();
-
-        getPromise(modalLooker)
-          .then((looker: ImaVidLooker) => {
-            if (looker) {
-              unsubscribe = looker.subscribeToState(key, (stateValue) => {
-                setSelf(stateValue);
-              });
-            }
-          })
-          .catch((e) => {
-            console.error(e);
-          });
-      });
-
-      return () => {
-        unsubscribe();
-      };
-    },
-  ],
-});
-
 export const groupStatistics = atomFamily<"group" | "slice", boolean>({
   key: "groupStatistics",
   default: "slice",
-});
-
-export const dynamicGroupFields = selector<string[]>({
-  key: "dynamicGroupFields",
-  get: ({ get }) => {
-    const groups = get(
-      fieldPaths({
-        ftype: EMBEDDED_DOCUMENT_FIELD,
-        embeddedDocType: GROUP,
-        space: State.SPACE.SAMPLE,
-      })
-    );
-    const lists = get(
-      fieldPaths({ ftype: LIST_FIELD, space: State.SPACE.SAMPLE })
-    );
-    const primitives = get(
-      fieldPaths({ ftype: DYNAMIC_GROUP_FIELDS, space: State.SPACE.SAMPLE })
-    ).filter((path) => path !== "filepath" && path !== "id");
-
-    const filtered = primitives.filter(
-      (path) =>
-        lists.every((list) => !path.startsWith(list)) &&
-        groups.every(
-          (group) => path !== `${group}.id` && path !== `${group}.name`
-        )
-    );
-
-    return filtered;
-  },
 });
