+++ app/packages/state/src/recoil/modal.ts
@@ -6,17 +6,16 @@ import { graphQLSelector } from "recoil-relay";
 import { VariablesOf } from "relay-runtime";
 import { Nullable } from "vitest";
 import { ResponseFrom } from "../utils";
+import { imaVidLookerState, shouldRenderImaVidLooker } from "./dynamicGroups";
 import {
   activeModalSidebarSample,
   groupId,
   groupSlice,
   hasGroupSlices,
-  imaVidLookerState,
   modalGroupSlice,
   pinned3DSample,
   pinned3DSampleSlice,
   pinned3d,
-  shouldRenderImaVidLooker,
 } from "./groups";
 import { RelayEnvironmentKey } from "./relay";
 import { datasetName } from "./selectors";
